+++ packages/jest-jasmine2/src/setup_jest_globals.ts
@@ -12,7 +12,6 @@ import {
   addSerializer,
   buildSnapshotResolver,
 } from 'jest-snapshot';
-import type {Plugin} from 'pretty-format';
 import type {
   Attributes,
   default as JasmineSpec,
@@ -22,7 +21,10 @@ import type {
 export type SetupOptions = {
   config: Config.ProjectConfig;
   globalConfig: Config.GlobalConfig;
-  localRequire: (moduleName: string) => Plugin;
+  localRequire: <T = unknown>(
+    path: string,
+    applyInteropRequireDefault?: boolean,
+  ) => Promise<T>;
   testPath: string;
 };
 
@@ -96,7 +98,7 @@ export default async function setupJestGlobals({
   // Jest tests snapshotSerializers in order preceding built-in serializers.
   // Therefore, add in reverse because the last added is the first tested.
   for (let i = config.snapshotSerializers.length - 1; i >= 0; i--) {
-    addSerializer(localRequire(config.snapshotSerializers[i]));
+    addSerializer(await localRequire(config.snapshotSerializers[i]));
   }
 
   patchJasmine();