+++ packages/jest-snapshot/src/SnapshotResolver.ts
@@ -9,7 +9,6 @@ import * as path from 'path';
 import chalk = require('chalk');
 import {createTranspilingRequire} from '@jest/transform';
 import type {Config} from '@jest/types';
-import {interopRequireDefault} from 'jest-util';
 
 export type SnapshotResolver = {
   /** Resolves from `testPath` to snapshot path. */
@@ -28,13 +27,16 @@ export const isSnapshotPath = (path: string): boolean =>
 
 const cache = new Map<string, SnapshotResolver>();
 
-type LocalRequire = (module: string) => unknown;
+type LocalRequire<T = unknown> = (
+  path: string,
+  applyInteropRequireDefault?: boolean,
+) => Promise<T>;
 
 export const buildSnapshotResolver = async (
   config: Config.ProjectConfig,
-  localRequire: Promise<LocalRequire> | LocalRequire = createTranspilingRequire(
-    config,
-  ),
+  localRequire:
+    | Promise<LocalRequire<SnapshotResolver>>
+    | LocalRequire<SnapshotResolver> = createTranspilingRequire(config),
 ): Promise<SnapshotResolver> => {
   const key = config.rootDir;
 
@@ -48,7 +50,7 @@ export const buildSnapshotResolver = async (
 };
 
 async function createSnapshotResolver(
-  localRequire: LocalRequire,
+  localRequire: LocalRequire<SnapshotResolver>,
   snapshotResolverPath?: string | null,
 ): Promise<SnapshotResolver> {
   return typeof snapshotResolverPath === 'string'
@@ -81,11 +83,12 @@ function createDefaultSnapshotResolver(): SnapshotResolver {
 
 async function createCustomSnapshotResolver(
   snapshotResolverPath: string,
-  localRequire: LocalRequire,
+  localRequire: LocalRequire<SnapshotResolver>,
 ): Promise<SnapshotResolver> {
-  const custom: SnapshotResolver = interopRequireDefault(
-    await localRequire(snapshotResolverPath),
-  ).default;
+  const custom: SnapshotResolver = await localRequire(
+    snapshotResolverPath,
+    true,
+  );
 
   const keys: Array<[keyof SnapshotResolver, string]> = [
     ['resolveSnapshotPath', 'function'],